+++ attributes/default.rb
@@ -21,12 +21,12 @@ case node['platform']
 when 'ubuntu','debian'
   shell = '/bin/false'
   homedir = '/var/lib/redis'
-when 'centos','redhat','fedora','scientific','amazon','suse'
+when 'centos','redhat','scientific','amazon','suse'
   shell = '/bin/sh'
-  homedir = '/var/lib/redis' #this is necessary because selinux by default prevents the homedir from being managed in /var/lib/
+  homedir = '/var/lib/redis' 
 when 'fedora'
   shell = '/bin/sh'
-  homedir = '/home'
+  homedir = '/home' #this is necessary because selinux by default prevents the homedir from being managed in /var/lib/ 
 else
   shell = '/bin/sh'
   homedir = '/redis'
@@ -39,7 +39,7 @@ default['redisio']['safe_install'] = true
 default['redisio']['mirror'] = "https://redis.googlecode.com/files"
 default['redisio']['base_name'] = 'redis-'
 default['redisio']['artifact_type'] = 'tar.gz'
-default['redisio']['version'] = '2.6.4'
+default['redisio']['version'] = '2.6.9'
 default['redisio']['base_piddir'] = '/var/run/redis'
 
 #Default settings for all redis instances, these can be overridden on a per server basis in the 'servers' hash
@@ -50,10 +50,13 @@ default['redisio']['default_settings'] = {
   'shell'                  => shell,
   'systemuser'             => true,
   'configdir'              => '/etc/redis',
+  'name'                   => nil,
   'address'                => nil,
   'databases'              => '16',
   'backuptype'             => 'rdb',
   'datadir'                => '/var/lib/redis',
+  'unixsocket'             => nil,
+  'unixsocketperm'         => nil,
   'timeout'                => '0',
   'loglevel'               => 'verbose',
   'logfile'                => nil,